produce valid univariant enums
Instead of generating a enum like this:

```
pub enum Enum_CUipcMem_flags_enum {
    CU_IPC_MEM_LAZY_ENABLE_PEER_ACCESS = 1,
}
```

Which is a compiler error because repr can't be used with univariant enums.
See https://github.com/rust-lang/rust/issues/10292

We generate a two variant enum with a dummy variant.

```
pub enum Enum_CUipcMem_flags_enum {
    CU_IPC_MEM_LAZY_ENABLE_PEER_ACCESS = 1,
    __DUMMY,
}
```

Which is the recommend workaround to the previous compiler error.

closes #255