Update quasi to 0.21
This is not the latest version, because the latest version doesn't work
due to serde-rs/quasi#65.

The issue of keeping using 0.20 is that libbindgen depends on aster 0.28
while quasi depends on aster 0.27, which leads to duplicate versions of
aster dependency. Updating quasi to 0.21 brings aster to 0.29, which
also works for libbindgen.