ir: Try to get C++ manglings with the appropriate API first.
As pointed out in #653, clang 3.8 asserts when trying to mangle a C++
constructor or a destructor.

The following patch tries to use the `clang_Cursor_getCXXManglings` function
first.

This assumes that the last mangling we're interested in is the proper one, which
seems to be true looking at LLVM, and on trunk on my machine.