Don't run test expectations' layout tests with multiple libclang versions
Only generating bindings depends on which libclang version we are dealing
with. Running the test expectations' layout tests does not change between
libclang versions, so these were redundant tests.

Note that we have not been testing the expected bindings that differ across
libclang versions (the nested sub-directories aren't automatically picked up by
cargo) and this commit does not change that.