Produce a more useful error message when rustfmt can't be found. Fixes #1205
Prior to this change bindgen would simply print any error that occurred
while attempting to run rustfmt straight to stderr using fmt::Debug.
Combined with the fact that rustfmt is enabled by default now this meant that
if rustfmt was missing or not working a cryptic error would be printed.