Allow static and dynamic linking
Currently bindgen always uses clang-sys with the "runtime" feature -
that is, load libclang at runtime with dlopen (or similar)
at runtime.  This PR keeps this default, but also
- adds "static" to statically link libclang
- without either "runtime" or "static", link with the shared library

Many distributions don't ship with a static libclang, but linking with the dynamic
library will use normal ld.so mechanisms to define where the .so file should be found.
(Ditto for the Mac and Windows equivalents.)