Zero out padding in custom Default trait implementations (#2051)
* Zero out padding in custom Default trait implementations

Previously, we were using `std::mem::zeroed()` which unfortunately does
not necessarily zero out padding. It'd be better if the padding is
zeroed out because some libraries are sensitive to non-zero'd out bytes,
especially when forward/backward compatability is involved.

This commit ensures all bytes are zeroed out in custom Default trait
implementations.