Don't assume that an inner class declaration always immediately yields a complete type.
It might not if we had to avoid recursion when processing types. Detect that
case and bail out. This bug was being masked by the fact that we didn't always
find definitions for the recursion check and so it didn't trigger, but now that
this check is more reliable we have to be careful in more places.

The test case was reduced from the GCC STL allocator definition.